Linebreak processing in MathML in Presentation XML: https://github.com/metanorma/metanorma-standoc/issues/111
